requirements.txt
vosk==0.3.45
pyaudio==0.2.14
pyttsx3==2.90
pygame==2.5.2
pystray==0.19.5
Pillow==10.4.0
numpy
